requirements.txt
asgiref==3.4.1           
certifi==2021.5.30            
django==3.2.5             
gunicorn==20.1.0             
pip==21.1.3                              
pytz==2021.1            
setuptools==49.6.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
wheel==0.36.2
whitenoise==5.3.0
wincertstore==0.2
Werkzeug==2.0.1
pymongo==3.12.0
djangorestframework==3.12.4
